Split IAPManager into app/tunnel .shared
