+++ Passepartout/Shared/Shared+Tunnel.swift
@@ -1,0 +1,50 @@
+//
+//  Shared+Tunnel.swift
+//  Passepartout
+//
+//  Created by Davide De Rosa on 11/14/24.
+//  Copyright (c) 2024 Davide De Rosa. All rights reserved.
+//
+//  https://github.com/passepartoutvpn
+//
+//  This file is part of Passepartout.
+//
+//  Passepartout is free software: you can redistribute it and/or modify
+//  it under the terms of the GNU General Public License as published by
+//  the Free Software Foundation, either version 3 of the License, or
+//  (at your option) any later version.
+//
+//  Passepartout is distributed in the hope that it will be useful,
+//  but WITHOUT ANY WARRANTY; without even the implied warranty of
+//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//  GNU General Public License for more details.
+//
+//  You should have received a copy of the GNU General Public License
+//  along with Passepartout.  If not, see <http://www.gnu.org/licenses/>.
+//
+
+import CommonLibrary
+import CommonUtils
+import Foundation
+
+extension IAPManager {
+    static let sharedForTunnel: IAPManager = {
+        IAPManager(
+            inAppHelper: Configuration.IAPManager.inAppHelper,
+            receiptReader: Configuration.IAPManager.tunnelReceiptReader,
+            productsAtBuild: Configuration.IAPManager.productsAtBuild
+        )
+    }()
+}
+
+private extension Configuration.IAPManager {
+
+    @MainActor
+    static var tunnelReceiptReader: AppReceiptReader {
+        // FIXME: #869, fallback release receipt is not available in Tunnel
+        fatalError()
+    }
+
+//    static var appGroupReceiptURLs: [URL] {
+//    }
+}
